feat: promote IRCd + VirtualClient to public embedding API (9.6.0, #22)
Closes #22. Unblocks culture#308 Phase A2-Bridge: culture's bot module
shrinks from a 231-LOC near-line-for-line copy of agentirc's internal
VirtualClient to a thin wrapper around the public class, and
culture/cli/server.py:_run_server can construct an agentirc.ircd.IRCd
in-process instead of going through the agentirc serve subprocess.

Promotions (additive-only, no behaviour change, same wire surface as 9.5.0):

- agentirc.virtual_client.VirtualClient is now public, semver-tracked.
  File moved from agentirc/_internal/virtual_client.py with no edits to
  the class body. The legacy import path resolves via a deprecation
  re-export module that emits DeprecationWarning; removal scheduled for
  10.0.0. Five in-tree call sites (ircd.py, server_link.py, channel.py
  x2, skills/rooms.py) retargeted at the public path.

- agentirc.ircd.IRCd is now public, semver-tracked. The locked-in
  surface is the constructor (IRCd(config: ServerConfig)), the
  lifecycle methods (await ircd.start() / stop()), await
  ircd.emit_event(event), and the subscription_registry / clients /
  channels / config / system_client attributes. Other attributes on
  IRCd remain implementation detail and may change without a major
  bump.

Documentation:

- docs/api-stability.md: new "Embedding agentirc in-process" section
  with worked example (~25 LOC), member-by-member breakdown of the
  IRCd public surface, VirtualClient signature reference, and 9.6.0
  versioning history row. agentirc.ircd removed from the
  internal-modules paragraph.

- CHANGELOG.md: new [9.6.0] - 2026-05-02 block under Added /
  Deprecated / Notes.

- CLAUDE.md: current-state header bumped to 9.6.0; public-API table
  expanded from 3 modules to 5 (now includes agentirc.ircd and
  agentirc.virtual_client).

Tests:

- tests/test_api_stability_embedding.py: 4 tests locking in the
  canonical public import paths, the deprecation shim's behaviour,
  and the named public members of IRCd and VirtualClient.

- Full suite: 410 passed (was 406; +4 from this file). pytest -n auto
  in ~30s.

Citation manifest:

- culture-virtual-client target moved from
  agentirc/_internal/virtual_client.py to agentirc/virtual_client.py.
- New agentirc-internal-virtual-client-shim entry (status: synthesize)
  for the deprecation re-export.

Out of scope: federation interop sniff (issue's optional ask 3 — defer
to a 9.5.x patch if culture asks); agentirc.host.Host facade (adds
surface area for no immediate benefit); Channel/Client/RemoteClient
promotion (not requested; embedders read ircd.channels and ircd.clients
as opaque mappings).
